commit script javítása

+++ commit.sh
@@ -5,26 +5,29 @@ m=$1
 if [[ -z '$m' ]]; then
     echo 'no commit message given'
 else
-    flutter pub get
+    current=$(git rev-parse --abbrev-ref HEAD)
+    if [[ $current == "master" ]]; then
+        echo {git branch | grep \* | cut -d ' ' -f2}
 
-    git add -A
-    git commit -a -m "$m"
-    echo 'rsync'
-    rsync -av --exclude='.git*' --exclude='README.md' ./ ../.szivacs-tmp/
-    echo 'rsyncend'
+        git checkout master
 
-    git checkout master
+        git add -A
+        git commit -a -m "$m"
+        rsync -av --exclude='.git*' --exclude='README.md' ./ ../.szivacs-tmp/
 
-    echo 'rsync'
-    rsync -av --exclude='.git*' --exclude='README.md' ../.szivacs-tmp/ ./
-    echo 'rsyncend'
-    sh to_foss.sh
-    flutter pub get
-    git add -A
-    git commit -a -m "$m"
+        git checkout foss
 
-    git checkout play-edition
+        rsync -av --exclude='.git*' --exclude='README.md' ../.szivacs-tmp/ ./
+        sh to_foss.sh
+        flutter pub get
+        git add -A
+        git commit -a -m "$m"
 
-    rm -rf ../.szivacs-tmp/
-    flutter pub get
+        git checkout master
+
+        rm -rf ../.szivacs-tmp/
+        flutter pub get
+    else
+        echo "Please use the master branch for committing changes!"
+    fi
 fi
